+++ src/Console/ModelsCommand.php
@@ -263,6 +263,7 @@ class ModelsCommand extends Command
             }
             $this->properties = [];
             $this->methods = [];
+            $this->foreignKeyConstraintsColumns = [];
             if (class_exists($name)) {
                 try {
                     // handle abstract classes, interfaces, ...
@@ -495,6 +496,10 @@ class ModelsCommand extends Command
         $schema = $model->getConnection()->getSchemaBuilder();
         $columns = $schema->getColumns($table);
         $platformName = $model->getConnection()->getConfig('driver');
+
+        if (strpos($table, '.')) {
+            [$database, $table] = explode('.', $table);
+        }
 
         if (!$columns) {
             return;
@@ -872,7 +877,12 @@ class ModelsCommand extends Command
 
     public function unsetMethod($name)
     {
-        unset($this->methods[strtolower($name)]);
+        foreach ($this->methods as $k => $v) {
+            if (strtolower($k) === strtolower($name)) {
+                unset($this->methods[$k]);
+                return;
+            }
+        }
     }
 
     public function getMethodType(Model $model, string $classType)